add count for Chrisy's eggs
Text only changes that will add a could of the clutches of eggs laid by
Christy. No new variables or mechanics.

Note: first attempt to provide a change for the game and this is
hiccup's original work. If there is any other preferred mechanism for
this please let me know, but it was not clear in the forum