bug fix in garrett sex scene progression
If a player has sex with Garrett only as a pure gryphon male, garrett_handjob_2 will be presented no matter how many times the player has sex with Garrett, and the hp of Garrett will never progress from 1.  In addition, garrett_handjob_1 has content for pure male gryphons, but it is not currently possible to ever get that scene as a pure male gryphon.

This code change should fix both of these issues.