Toy Store event rewrite, kinky perk alias
Rewrote the Toy Store event with new descriptions and a unique sex scene on loss.

Dys was instrumental with the Toy Store revamp, helping write the new sex scene.

Also added an alias for checking the "Kinky" feat, as it is somewhat commonly called and makes it easier to add scenes for it.