General code and writing cleanup
Consolidated all park events into a single file, but as separate, searchable events.

Changed outlying uses of 'normally' to 'usually' when definining a default value. 'normally' is a valid alternative but non-standard for Inform - the documentation doesn't reference it much. This change is mainly to make it easier to onboard new writers with more consistent syntax.

Fixes various typos and added missing commas where needed.