Fix Abandoned Truck event not always being reset
### Purpose of the PR
This is a *potential fix* for the issue [EJSteff reported over at Discord](https://discordapp.com/channels/333559467218173953/344610529223901184/600125514069442602): Under certain circumstances it is possible, that the 'Abandoned Truck'-event doesn't get unresolved after getting the quest to look for it from Doctor Medea. However I couldn't repeat this a neither does EjSteff. Hence, why this is a *potential fix*.

### Changes
Instead of spamming the code with
```inform7
now Abandoned Truck is active;
now Abandoned Truck is not resolved;
```
I've added a function that resets it if needed aka `if Medeaget is 1`. The function is called at the end of each talk with Medea aka just in time.

### Notes
Did a couple tests and so far the truck gets reset and I can finish Medeas quest.

### Discord handle
Stadler#3007