Wyverns should ovi
I've always wondered, why wyverns give birth instead of laying eggs
unless the ovi setting is set to **Always Ovi**, while others, like
hawkman female and blue gryphon herm do.
Is this a bug or is it intended?
Well, instead of wating for an answer on Discord I've decided to file a
PR here. Its a simple search and replace [m?impregchance] with
[m?ovichance] aka `s/\[(m)?impregchance\]/[$1ovichance]/i`

I'll leave it up to you to decide wheter to merge or reject the PR.