Made sync.bat to work in localized environments
### Purpose of the PR
`sync.bat` didn't work in localized environments.

### Changes
`find "Symbolic Link"` doesn't match in non-english environments. Removed that, since its not necessary: `fsutil reparsepoint query` errors in 2 cases: 1.: The file/folder is not a symlink and 2.: user is not an admin. Since the script ensures an elevated (admin) commandline case 2 won't happen, so we're good to go without matching.

### Notes
- Tested and working on my localized machine.
- No recompile needed.
- Can repeatedly run that script without messing things up.