First version of the Table of GameEndings implementation
Ideally every game ending should be handled in a single table which gets iterated 'when play ends:'.
The plan is to reduce every 'when play ends:' into a single one, which iterates through that table which is ordered by priority.
In the example, I use:
```inform7
if Player imprisoned of TheEnd is true:
	stop the action;
```
This should be replaced with `break;` later on because `stop the action;` ends the whole `when play ends:` which is usually unintended.
e. g. the advert for FS Multiplayer won't show up after a `stop the action;` in any `when play ends:`.