Using turn the Player into where applicable
I've searched through the codebase to find any place, where the player is fully TFed into another 'infection' and applied `turn the Player into a "...";` to it, where applicable.
With this, its made sure, that `SpeciesName` are being set corrently and it calls `attributeinfect "...";` so the size change 'n stuff is done as well.
As a bonus it should fix many borked children due to the lack of updating `SpeciesName`s and other stuff properly.