Excluded Sarah's Epilogue, when the player succumbed and has left
If the player left everything behind and has succumbed to the infection, she or he won't return to the bunker to screw up Sarah and spent the rest of her/his life with her.

Though I could think of one variation of this (if a writer wants to do the scene for this):
If the player hit the "Timothy's Breeding Gryphon" she might turn Sarah into another breeder and take her back to Qytat.