Fix Blue Gryphon Herm full TF for Guard Gryphon event
Accepting the full Gryphon TF from the Guard Gryphon missed the SpeciesName's and the gender wasn't 'changed' as well as the breasts weren't growing, so that for example a flat breasted male human became a flat breasted male blue gryphon.
Fixed that with a turn the Player into and following the sex change and breast change rules.