Adding new walkin events for Ava and activating the newsstand raid event in the High Rise District.
Adds) Newsstand Raid to the High Rise District, it can not be failed and will lead to Ava finishing her art project at a later date.

Adds) Art supplies: A follow up walk in event at the Last Chance Auto Body Shop, mostly plot/lore based and drops some hints about Drake and Ava’s past.

Adds) Margay Breeding: A preg focused walk in event that kicks off Ava’s preg quest. To view it the player must be pregnant, a pureblood margay taur, female, be size five, have met Kathy, not be in the rebuked state, and Ava must not have the trait "Propositioned".

Adds) Mirror mirror on Ava’s wall: A follow up event that gives the player one last chance to bail on Ava’s preg quest. Otherwise, this is mostly an RP walk in for people that want to give their toons a bit of flavor or express some opinion on being knocked up. To view it the player must be pregnant, a pureblood margay taur, female, be size five, not be in the rebuked state, and Ava must not have the trait "Propositioned".

Adds) Penny for your thoughts: Another RP walk in event, no sex, just plot/lore. To view it the player must be pregnant, a pureblood margay taur, female, be size five, not be in the rebuked state, and Ava must not have the trait "Propositioned".

Adds) Broodmother Threshold: Lewd walk in event for the player, depending on their choice in the previous events upon completion this will start Ava’s Fpreg quest. To view it the player must be pregnant, a pureblood margay taur, female, be size five, not be in the rebuked state, and Ava must not have the trait "Propositioned".

Adds) The trait Violence of action: Obtained by the player handling the Newsstand raid in an aggressive manner. Collecting this trait and two out of five others will lead to one of the player's offspring coming out as vore/UB preds.

Fixes) Drake and Kathy no longer have ‘bards’ on their tongues/cock and some minor typos have been corrected. The stable Warp rift linking the Sniper’s Perch and Parking lot has been sealed.