adds a new area to the State Fair and a new secret character
The player may visit an empty and abandoned side of the fair.
If the player is level 20 or higher and is touched by madness, they may encounter a new creature.
Depending on how the player interacts with this creature, it will either become an ally, a future threat, or remain as is.
One of the things she's meant to help with is the huge amount of junk late game players have, and instead of trashing it all, some of it can be recycled into something better or useful. The list of items that can be recycled is small for now, but I hope to add more as I work on her.

For those who don't feel like grinding or just want to test her content, here's the debug codes:
"SummonTheLimbo" or "SummonCyn" : skips the first events and places Cyn in the attic of the library.
"Give Echos" : gives 1000 of this NPC's specific currency and a debug item that gives or removes touched by madness when used.